backend/requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
PyMySQL==1.1.1
python-dotenv==1.0.1
email-validator==2.2.0
pytest==8.3.5
gunicorn==23.0.0
